Cleaning up
Looking to implement the calendar's read and write functionality, I found that it revolves around the Event class. This class is probably unnecessary if all events will be stored on a text file and not a database. So I removed references to that class in favor of using only the Calendar class to manipulate text files.

I will leave the Event class unchanged, but with a comment at the top in case it would be required again.